ml-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy==1.26.4
scikit-learn==1.4.2
python-dotenv==1.0.0
joblib==1.3.2
